# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Urja Meter Ops API",
    "description": "A clean, documented REST API in front of the Urja Meter Ops utility portal. See /docs for interactive exploration, and the project's PROTOCOL.md for how the underlying portal actually works.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/meters": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "List / search / filter meters",
        "description": "List meters with filters the portal's own UI can't combine (it only\noffers a single free-text box). Backed by our in-memory index, not a\nlive portal call \u2014 see PROTOCOL.md / README for why.",
        "operationId": "list_meters_api_v1_meters_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text match against meter ID or serial number.",
              "title": "Q"
            },
            "description": "Free-text match against meter ID or serial number."
          },
          {
            "name": "install_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Installed | Faulty | Decommissioned",
              "title": "Install Status"
            },
            "description": "Installed | Faulty | Decommissioned"
          },
          {
            "name": "phase_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "single | three",
              "title": "Phase Type"
            },
            "description": "single | three"
          },
          {
            "name": "make",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Make"
            }
          },
          {
            "name": "install_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "CT Operated | Whole Current",
              "title": "Install Type"
            },
            "description": "CT Operated | Whole Current"
          },
          {
            "name": "dt_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to meters under this distribution transformer.",
              "title": "Dt Code"
            },
            "description": "Filter to meters under this distribution transformer."
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Zone code, e.g. Z-01.",
              "title": "Zone"
            },
            "description": "Zone code, e.g. Z-01."
          },
          {
            "name": "circle",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Circle"
            }
          },
          {
            "name": "division",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Division"
            }
          },
          {
            "name": "subdivision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subdivision"
            }
          },
          {
            "name": "substation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Substation"
            }
          },
          {
            "name": "feeder",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feeder"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMeters"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/near": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "Meters near a lat/lng point",
        "description": "Not something the portal can answer at all \u2014 it has no spatial\nquery. Straight-line (haversine) distance, sorted nearest-first.",
        "operationId": "meters_near_api_v1_meters_near_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lat"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lng"
            }
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Omit for no distance cap.",
              "title": "Radius Km"
            },
            "description": "Omit for no distance cap."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NearbyMeter"
                  },
                  "title": "Response Meters Near Api V1 Meters Near Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "Get a single meter's nameplate, hierarchy and geo",
        "operationId": "get_meter_api_v1_meters__meter_id__get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meter"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}/consumption": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "Consumption readings for a meter, with a computed usage summary",
        "description": "Raw readings from the portal are cumulative register values, not\ninterval usage \u2014 this endpoint returns both the raw readings *and* a\n`summary` with derived consumption (last - first) and the inferred\nreading resolution, since the portal's own granularity is inconsistent\nacross meters (see PROTOCOL.md).",
        "operationId": "get_meter_consumption_api_v1_meters__meter_id__consumption_get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterConsumption"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transformers": {
      "get": {
        "tags": [
          "transformers"
        ],
        "summary": "List distribution transformers (DTs)",
        "description": "Each DT is enriched with `meter_count` / `status_breakdown` computed\nfrom the meter dataset \u2014 the portal's own DT listing doesn't include\neither.",
        "operationId": "list_transformers_api_v1_transformers_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTransformers"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transformers/{code}": {
      "get": {
        "tags": [
          "transformers"
        ],
        "summary": "Get a single transformer",
        "operationId": "get_transformer_api_v1_transformers__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transformer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConsumptionReading": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Naive local timestamp as reported by the portal (no timezone given \u2014 see PROTOCOL.md)."
          },
          "kwh": {
            "type": "number",
            "title": "Kwh",
            "description": "Cumulative active-energy register reading, not an interval delta."
          },
          "kvah": {
            "type": "number",
            "title": "Kvah",
            "description": "Cumulative apparent-energy register reading, not an interval delta."
          },
          "volt_r": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volt R",
            "description": "R-phase voltage at the time of the reading."
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "kwh",
          "kvah"
        ],
        "title": "ConsumptionReading"
      },
      "ConsumptionResolution": {
        "type": "string",
        "enum": [
          "sub_hourly",
          "daily",
          "unknown"
        ],
        "title": "ConsumptionResolution"
      },
      "ConsumptionSummary": {
        "properties": {
          "reading_count": {
            "type": "integer",
            "title": "Reading Count"
          },
          "resolution": {
            "$ref": "#/components/schemas/ConsumptionResolution",
            "description": "Inferred from the median gap between readings. The portal returns wildly different granularities per meter for no metadata-visible reason (see PROTOCOL.md) so callers should not assume a fixed interval."
          },
          "first_timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Timestamp"
          },
          "last_timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Timestamp"
          },
          "kwh_consumed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kwh Consumed",
            "description": "last.kwh - first.kwh over the returned window."
          },
          "kvah_consumed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kvah Consumed"
          },
          "avg_volt_r": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Volt R"
          },
          "min_volt_r": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Volt R"
          },
          "max_volt_r": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Volt R"
          }
        },
        "type": "object",
        "required": [
          "reading_count",
          "resolution"
        ],
        "title": "ConsumptionSummary"
      },
      "GeoPoint": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "title": "GeoPoint"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HierarchyRef": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "HierarchyRef"
      },
...
      "InstallStatus": {
        "type": "string",
        "enum": [
          "Installed",
          "Faulty",
          "Decommissioned"
        ],
        "title": "InstallStatus"
      },
      "InstallType": {
        "type": "string",
        "enum": [
          "CT Operated",
          "Whole Current"
        ],
        "title": "InstallType"
      },
      "Meter": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "serial_no": {
            "type": "string",
            "title": "Serial No"
          },
          "make": {
            "type": "string",
            "title": "Make"
          },
          "phase_type": {
            "$ref": "#/components/schemas/PhaseType"
          },
          "install_status": {
            "$ref": "#/components/schemas/InstallStatus"
          },
          "dt_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dt Code"
          },
          "install_type": {
            "$ref": "#/components/schemas/InstallType"
          },
          "build": {
            "type": "string",
            "title": "Build"
          },
          "hierarchy": {
            "$ref": "#/components/schemas/MeterHierarchy"
          },
          "geo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoPoint"
              },
              {
                "type": "null"
              }
            ]
          },
          "hierarchy_complete": {
            "type": "boolean",
            "title": "Hierarchy Complete",
            "description": "False if any level of `hierarchy` was missing in the source data."
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "install_status",
          "install_type",
          "build",
          "hierarchy",
          "hierarchy_complete"
        ],
        "title": "Meter",
        "description": "Full meter record, as reconstructed from the portal's bulk export."
      },
      "MeterConsumption": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "summary": {
            "$ref": "#/components/schemas/ConsumptionSummary"
          },
          "readings": {
            "items": {
              "$ref": "#/components/schemas/ConsumptionReading"
            },
            "type": "array",
            "title": "Readings"
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "summary",
          "readings"
        ],
        "title": "MeterConsumption"
      },
      "MeterHierarchy": {
        "properties": {
          "zone": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "circle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "division": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "subdivision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "substation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "feeder": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "dt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "MeterHierarchy",
        "description": "The network path from a meter up to its zone.\n\nAny level can legitimately be `None` \u2014 see PROTOCOL.md: ~5% of meters in\nthe source data are missing one or more levels (empty string in the\nportal's own payload). We normalise \"\" to None rather than pretending\nthe utility has no gaps."
      },
      "MeterSummary": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "serial_no": {
            "type": "string",
            "title": "Serial No"
          },
          "make": {
            "type": "string",
            "title": "Make"
          },
          "phase_type": {
            "$ref": "#/components/schemas/PhaseType"
          },
          "install_status": {
            "$ref": "#/components/schemas/InstallStatus"
          },
          "dt_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dt Code"
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "install_status"
        ],
        "title": "MeterSummary",
        "description": "Row shape used in list/search results \u2014 cheap fields only."
      },
      "NearbyMeter": {
        "properties": {
          "meter": {
            "$ref": "#/components/schemas/MeterSummary"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          }
        },
        "type": "object",
        "required": [
          "meter",
          "distance_km"
        ],
        "title": "NearbyMeter"
      },
      "PaginatedMeters": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MeterSummary"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "PaginatedMeters"
      },
      "PaginatedTransformers": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Transformer"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "PaginatedTransformers"
      },
      "PhaseType": {
        "type": "string",
        "enum": [
          "single",
          "three"
        ],
        "title": "PhaseType"
      },
      "Transformer": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "feeder_code": {
            "type": "string",
            "title": "Feeder Code"
          },
          "capacity_kva": {
            "type": "number",
            "title": "Capacity Kva"
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count",
            "default": 0
          },
          "status_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Status Breakdown"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "feeder_code",
          "capacity_kva"
        ],
        "title": "Transformer",
        "description": "A distribution transformer (DT)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}